DEV: move building constructor to wasm engine module
