+++ enigma-core/enclave/src/lib.rs
@@ -51,7 +51,6 @@ use enigma_tools_t::common::{errors_t::{EnclaveError, EnclaveError::*, FailedTas
 use enigma_tools_m::utils::{EthereumAddress, LockExpectMutex};
 use enigma_tools_t::{build_arguments_g::*, quote_t, storage_t, esgx::ocalls_t};
 use enigma_types::{traits::SliceCPtr, EnclaveReturn, ExecuteResult, Hash256, ContractAddress, PubKey, ResultStatus, RawPointer, DhKey};
-use wasm_utils::{build, SourceTarget};
 
 use sgx_types::*;
 use std::{mem, ptr, slice, str};
@@ -335,7 +334,7 @@ unsafe fn ecall_execute_internal(pre_execution_data: &mut Vec<Box<[u8]>>, byteco
              map_err(|e| {FailedTaskError(InputError{ message: format!("{}", e) })})?;
 
     let state_key = km_t::get_state_key(address)?;
-    let mut engine = WasmEngine::new(&bytecode, gas_limit, decrypted_args.clone(), pre_execution_state.clone(), function_name, state_key)?;
+    let mut engine = WasmEngine::new_compute(&bytecode, gas_limit, decrypted_args.clone(), pre_execution_state.clone(), function_name, state_key)?;
     engine.compute()?;
     let exec_res = engine.into_result()?;
     let delta_hash = save_enc_delta(db_ptr, &exec_res.state_delta)?;
@@ -367,42 +366,7 @@ unsafe fn ecall_execute_internal(pre_execution_data: &mut Vec<Box<[u8]>>, byteco
     Ok(())
 }
 
-/// Builds Wasm code for contract deployment from the Wasm contract.
-/// Gets byte vector with Wasm code.
-/// Created code contains one function `call`, which invokes `deploy`.
-/// `deploy` invokes the contract constructor from `wasm_code` and returns the bytecode to be deployed
-/// Writes created code to a file constructor.wasm in a current directory.
-/// This code is based on https://github.com/paritytech/wasm-utils/blob/master/cli/build/main.rs#L68
-/// The parameters' values to build function are default parameters as they appear in the original code.
-pub fn build_constructor(wasm_code: &[u8]) -> Result<Vec<u8>, EnclaveError> {
-    let module = parity_wasm::deserialize_buffer(wasm_code)?;
 
-    let (module, ctor_module) = match build(
-        module,
-        SourceTarget::Unknown,
-        None,
-        &Vec::new(),
-        false,
-        "49152".parse().expect("New stack size is not valid u32"),
-        false,
-    ) {
-        Ok(v) => v,
-        Err(e) => panic!("build_constructor: {:?}", e), // TODO: Return error
-    };
-
-    let result;
-
-    if let Some(ctor_module) = ctor_module {
-        result = parity_wasm::serialize(ctor_module); /*.map_err(Error::Encoding)*/
-    } else {
-        result = parity_wasm::serialize(module); /*.map_err(Error::Encoding)*/
-    }
-
-    match result {
-        Ok(v) => Ok(v),
-        Err(e) => panic!("build_constructor: {:?}", e), // TODO: Return Error
-    }
-}
 
 unsafe fn ecall_deploy_internal(pre_execution_data: &mut Vec<Box<[u8]>>, bytecode: &[u8], constructor: &[u8], args: &[u8],
                                 address: ContractAddress, user_key: &PubKey, io_key: &DhKey,
@@ -413,14 +377,13 @@ unsafe fn ecall_deploy_internal(pre_execution_data: &mut Vec<Box<[u8]>>, bytecod
     let inputs_hash = enigma_crypto::hash::prepare_hash_multiple(&[constructor, args, &pre_code_hash[..], user_key][..]).keccak256();
     pre_execution_data.push(Box::new(*inputs_hash));
 
-    let deploy_bytecode = build_constructor(bytecode)?;
     let (decrypted_args, function_name) = decrypt_inputs(constructor, args, io_key).
         map_err(|e| {FailedTaskError(InputError{ message: format!("{}", e) })})?;
 
     let state = ContractState::new(address);
 
     let state_key = km_t::get_state_key(address)?;
-    let mut engine = WasmEngine::new(&deploy_bytecode, gas_limit, decrypted_args.clone(), state, function_name, state_key)?;
+    let mut engine = WasmEngine::new_deploy(bytecode, gas_limit, decrypted_args.clone(), state, function_name, state_key)?;
     engine.deploy()?;
     let exec_res = engine.into_result()?;
