+++ enigma-runtime-t/src/gas.rs
@@ -1,9 +1,5 @@
-/// This module builds Wasm code for contract deployment from the Wasm contract.
-/// The contract should be written in rust and then compiled to Wasm with wasm32-unknown-unknown target.
-/// The code is based on Parity wasm_utils::cli.
-extern crate pwasm_utils;
 
-pub use self::pwasm_utils::{inject_gas_counter, rules};
+pub use pwasm_utils::{inject_gas_counter, rules};
 
 /// Wasm cost table
 pub struct WasmCosts {
